+++ frontend/src/components/CharacterClass.vue
