+++ frontend/src/components/HitPoints.vue
@@ -1,33 +1,61 @@
 <template>
   <div>
-    <v-btn @click="hitpointDialog=true" round flat>
+    <v-btn
+      @click="hitpointDialog=true"
+      round
+      flat
+    >
       <v-layout justify-space-around align-center>
         <v-icon>mdi-heart</v-icon>
-        <span>{{hitpoints}} / {{maxHitpoints}}</span>
+        <span>{{ character.hitpoints }} / {{ character.maxHitpoints }}</span>
       </v-layout>
     </v-btn>
     <!-- Hit Point Counter md-heart -->
     <!-- Hit Point Roll md-refresh -->
     <!-- Receive Attack Dialog md-sword -->
-    <v-dialog v-model="hitpointDialog" max-width="500">
+    <v-dialog
+      v-model="hitpointDialog"
+      max-width="500"
+    >
       <v-card>
         <v-card-title>
           <h2>Health</h2>
         </v-card-title>
         <v-card-text>
-          <v-layout align-center justify-center row fill-height>
+          <v-layout
+            align-center
+            justify-center
+            row
+            fill-height
+          >
             <v-tooltip top>
               <!-- TODO Replace with broken shield icon -->
-              <v-btn large icon flat slot="activator" @click="hurt">
+              <v-btn
+                large
+                icon
+                flat
+                slot="activator"
+                @click="hurt"
+              >
                 <v-icon>mdi-sword</v-icon>
               </v-btn>
               <span>Take Damage</span>
             </v-tooltip>
             <v-flex xs1>
-              <v-text-field single-line :rules="[mustBeNum, minNum]" v-model="offset"/>
+              <v-text-field
+                single-line
+                :rules="[mustBeNum, minNum]"
+                v-model="offset"
+              />
             </v-flex>
             <v-tooltip top>
-              <v-btn large icon flat slot="activator" @click="heal">
+              <v-btn
+                large
+                icon
+                flat
+                slot="activator"
+                @click="heal"
+              >
                 <v-icon>mdi-medical-bag</v-icon>
               </v-btn>
               <span>Heal</span>
@@ -35,7 +63,12 @@
           </v-layout>
         </v-card-text>
         <v-card-text>
-          <v-layout align-center justify-space-between row fill-height>
+          <v-layout
+            align-center
+            justify-space-between
+            row
+            fill-height
+          >
             <h3>Current Health :</h3>
             <v-flex xs2>
               <v-text-field
@@ -47,7 +80,11 @@
               />
             </v-flex>
           </v-layout>
-          <v-layout align-center justify-space-between row>
+          <v-layout
+            align-center
+            justify-space-between
+            row
+          >
             <h3>Maximum Health :</h3>
             <v-flex xs2>
               <v-text-field
@@ -63,13 +100,23 @@
         <v-card-text>
           <v-layout justify-end row>
             <v-tooltip top>
-              <v-btn icon flat slot="activator" @click="getHealth(true)">
+              <v-btn
+                icon
+                flat
+                slot="activator"
+                @click="getHealth(true)"
+              >
                 <v-icon>mdi-dice-multiple</v-icon>
               </v-btn>
               <span>Roll Health</span>
             </v-tooltip>
             <v-tooltip top>
-              <v-btn icon flat slot="activator" @click="getHealth(false)">
+              <v-btn
+                icon
+                flat
+                slot="activator"
+                @click="getHealth(false)"
+              >
                 <v-icon>mdi-heart-half-full</v-icon>
               </v-btn>
               <span>Take Avg for Health</span>
@@ -82,24 +129,23 @@
 </template>
 
 <script>
+import { hitDice } from '../common/constants'
+
 export default {
-  props: ['charIndex'],
+  props: ['character'],
+
   mounted() {
     this.getHealth(false)
     this.offset = this.maxHitpoints
     this.heal()
     this.offset = '0'
   },
+
   computed: {
-    character() {
-      return this.$store.state.characters[this.charIndex]
-    },
     deathThrows() {
       return this.character.deathThrows
     },
-    hitDice() {
-      return this.$store.state.hitDice
-    },
+
     hitpoints: {
       get() {
         return this.character.hitpoints
@@ -161,7 +207,7 @@ export default {
         for (let l = 0; l < this.character.classes[c].level; l++) {
           totalHealth += constitutionOffset
           let value = 0
-          const dice = this.hitDice[
+          const dice = hitDice[
             this.character.classes[c].classname.split(' ')[0]
           ]
           if (firstLevel) {
