+++ frontend/src/components/Tracker.vue
@@ -1,76 +1,9 @@
 <template>
   <v-card>
-    <!-- actions -->
-    <v-toolbar color="secondary" flat>
-      <v-tooltip bottom>
-        <v-btn
-          icon
-          flat
-          slot="activator"
-          color="primary"
-          @click="multiclass()"
-        >
-          <v-icon>add_circle_outline</v-icon>
-        </v-btn>
-        <span>Multiclass</span>
-      </v-tooltip>
-      <v-tooltip bottom>
-        <v-btn
-          color="primary"
-          flat
-          slot="activator"
-          icon
-          @click="castSpell()"
-        >
-          <v-icon>mdi-auto-fix</v-icon>
-        </v-btn>
-        <span>Cast Spell</span>
-      </v-tooltip>
-      <v-tooltip bottom>
-        <v-btn
-          color="primary"
-          icon
-          flat
-          slot="activator"
-          @click="longRest()"
-        >
-          <v-icon>mdi-sleep</v-icon>
-        </v-btn>
-        <span>Long Rest</span>
-      </v-tooltip>
-      <v-tooltip bottom>
-        <v-btn
-          color="primary"
-          icon
-          flat
-          slot="activator"
-          @click="shortRest()"
-        >
-          <v-icon>mdi-bell-sleep</v-icon>
-        </v-btn>
-        <span>Short Rest</span>
-      </v-tooltip>
-      <v-tooltip bottom>
-        <v-btn
-          :disabled="!character.concentrating"
-          @click="concentrationDialog=true"
-          color="primary"
-          icon
-          flat
-          slot="activator"
-        >
-          <v-icon>remove_red_eye</v-icon>
-        </v-btn>
-        <span
-          v-if="character.concentrating"
-        >Concentrating on {{ character.concentrating }}</span>
-        <span
-          v-if="!character.concentrating || character.concentrating === ''"
-        >Not currently concentrating</span>
-      </v-tooltip>
-      <v-spacer></v-spacer>
-      <h3>PROFICIENCY BONUS: +{{ character.proficiency }}</h3>
-    </v-toolbar>
+    <app-character-actions
+      :character="character"
+    />
+    <!-- flesh out and make collapsible -->
     <!-- character info -->
     <v-layout>
       <v-flex>
@@ -91,44 +24,18 @@
     <!-- dynamic state -->
     <v-layout column>
       <app-initiative :character="character"/>
-      <v-layout align-center row>
-        <!-- Hit Points -->
-        <hit-points :character="character"></hit-points>
-        <!-- End Hit Points -->
-      </v-layout>
+      <app-hit-points :character="character"/>
     </v-layout>
-    <death-throws
-      v-if="hitpoints <= 0 && maxHitpoints > 0"
+    <app-death-throws
+      v-if="character.hitpoints <= 0 && character.maxHitpoints > 0"
       :character="character"
     />
-    <v-dialog
-      v-model="concentrationDialog"
-      max-width="300"
-    >
-      <v-card>
-        <v-card-text>
-          <h2>Stop Concentrating on {{character.concentrating}}?</h2>
-        </v-card-text>
-        <v-layout column>
-          <v-btn
-            @click="concentrationDialog = false; stopConcentrating()"
-            flat
-            color="primary"
-          >Yes</v-btn>
-          <v-btn
-            @click="concentrationDialog = false;"
-            flat
-            color="primary"
-          >No</v-btn>
-        </v-layout>
-      </v-card>
-    </v-dialog>
     <!-- End "Buttons" -->
     <!-- Ability Scores -->
-    <ability-scores
+    <app-ability-scores
       :scores="character.abilityScores"
       :character="character"
-    ></ability-scores>
+    />
     <!-- End Ability Scores -->
     <v-card-text
       v-for="(characterClass, classindex) in character.classes"
@@ -136,13 +43,13 @@
     >
       <!-- TODO Class-specific stuff -->
 
-      <character-class
+      <app-character-class
         :characterClass="characterClass"
         :character="character"
         :classIndex="classindex"
-      ></character-class>
+      />
     </v-card-text>
-    <spell-cast
+    <app-spell-cast
       :character="character"
       ref="spellCast"
     />
@@ -150,12 +57,13 @@
 </template>
 
 <script>
-import AbilityScores from '@/components/AbilityScores'
-import Class from '@/components/Class'
-import SpellCast from '@/components/SpellCast'
-import HitPoints from '@/components/HitPoints'
+import CharacterActions from './CharacterActions'
+import AbilityScores from './AbilityScores'
+import CharacterClass from './CharacterClass'
+import SpellCast from './SpellCast'
+import HitPoints from './HitPoints'
 import DeathSavingThrows from './DeathSavingThrows'
-import Initiative from '@/components/Initiative'
+import Initiative from './Initiative'
 
 export default {
   name: 'tracker',
@@ -163,8 +71,9 @@ export default {
   props: ['character'],
 
   components: {
+    'app-character-actions': CharacterActions,
     'app-ability-scores': AbilityScores,
-    'app-character-class': Class,
+    'app-character-class': CharacterClass,
     'app-death-throws': DeathSavingThrows,
     'app-spell-cast': SpellCast,
     'app-hit-points': HitPoints,
@@ -175,106 +84,6 @@ export default {
     return {
       localName: this.character.name,
     }
-  },
-
-  computed: {
-    hitDice() {
-      return this.$store.state.hitDice
-    },
-
-    hitpoints() {
-      return this.character.hitpoints
-    },
-
-    maxHitpoints() {
-      return this.character.maxHitpoints
-    },
-
-    name: {
-      get() {
-        return this.character.name
-      },
-      set(state) {
-        this.$store.commit('changeName', {
-          index: this.index,
-          name: state,
-        })
-      },
-    },
-  },
-  data() {
-    return {
-      concentrationDialog: false,
-      shortRestDie: {},
-    }
-  },
-  methods: {
-    multiclass() {
-      this.$store.commit('multiclass', { index: this.index, classname: '' })
-    },
-    longRest() {
-      this.$store.commit('longRest', this.index)
-    },
-    resetCharacter() {
-      for (const c in this.character.classes) {
-        this.$store.commit('updateSlots', {
-          charIndex: this.index,
-          classIndex: c,
-        })
-      }
-    },
-    performShortRest() {
-      let restoredHealth = 0
-      for (const a in this.shortRestDie) {
-        restoredHealth += Math.floor(Math.random() * this.shortRestDie[a])
-      }
-      this.$store.commit('setHP', {
-        charIndex: this.index,
-        hitpoints: parseInt(this.hitpoints) + restoredHealth,
-      })
-      this.$store.commit('hideSnackbar')
-    },
-    shortRest() {
-      const acc = {}
-      for (const c in this.character.classes) {
-        if (
-          !this.hitDice.hasOwnProperty(
-            this.character.classes[c].classname.split(' ')[0],
-          )
-        ) {
-          continue
-        }
-        const hitDie = this.hitDice[
-          this.character.classes[c].classname.split(' ')[0]
-        ]
-        if (!acc.hasOwnProperty(hitDie)) {
-          acc[hitDie] = 0
-        }
-        acc[hitDie] += this.character.classes[c].level
-      }
-      let message = []
-      this.shortRestDie = []
-      for (const k in acc) {
-        message.push(acc[k].toString() + 'd' + k.toString())
-        this.shortRestDie.push.apply(
-          this.shortRestDie,
-          new Array(acc[k]).fill(k),
-        )
-      }
-      message = 'Restore ' + message.join(', ')
-      this.$store.commit('showSnackbar', {
-        color: 'green',
-        message,
-        func: this.performShortRest,
-        buttonMessage: 'Roll',
-      })
-    },
-    stopConcentrating() {
-      this.$store.commit('stopConcentrating', this.index)
-    },
-    castSpell() {
-      this.$refs.spellCast.spellPreflight()
-    },
   },
 }
 </script>
